Add files via upload
Commented out functions
Added new modules
